Use AWSStaticCredentialsProvider and DefaultAWSCredentialsProviderChain
Fixes #144
